refactor : 컨트롤러 로직 도메인으로 이동 리팩토링
